requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
opencv-python==4.8.0.74
pyzbar==0.1.8
requests==2.31.0
Werkzeug==2.3.7